board/requirements.txt
anyio==3.3.0
certifi==2021.5.30
colorzero==2.0
gpiozero==1.6.2
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httpcore==0.13.6
httpx==0.18.2
hyperframe==5.2.0
idna==3.2
pkg-resources==0.0.0
playsound==1.3.0
pycairo==1.20.1
pydub==0.25.1
python-dotenv==0.19.0
rfc3986==1.5.0
RPi.GPIO==0.7.0
simpleaudio==1.0.4
sniffio==1.2.0
typing-extensions==3.10.0.0
